Split Codex shim modules
