+++ crates/defra-agent-cli/src/commands/codex_shim/handlers/basic.rs
@@ -1,0 +1,198 @@
+use anyhow::Result;
+use codex_app_server_protocol as codex;
+use serde_json::json;
+
+use super::super::protocol::{
+    absolute_path, empty_rate_limits, initialize_result, model_summary, send_result,
+    send_typed_json_result,
+};
+use super::super::{Outbound, ShimState};
+
+pub(super) async fn handle_basic_request(
+    outbound: &Outbound,
+    state: &ShimState,
+    request: codex::ClientRequest,
+) -> Result<()> {
+    match request {
+        codex::ClientRequest::Initialize { request_id, .. } => {
+            send_typed_json_result::<codex::InitializeResponse>(
+                outbound,
+                request_id,
+                initialize_result(state),
+            )
+            .await
+        }
+        codex::ClientRequest::GetAccount { request_id, .. } => {
+            send_result(
+                outbound,
+                request_id,
+                codex::GetAccountResponse {
+                    account: Some(codex::Account::ApiKey {}),
+                    requires_openai_auth: false,
+                },
+            )
+            .await
+        }
+        codex::ClientRequest::GetAccountRateLimits { request_id, .. } => {
+            send_result(
+                outbound,
+                request_id,
+                codex::GetAccountRateLimitsResponse {
+                    rate_limits: empty_rate_limits(),
+                    rate_limits_by_limit_id: None,
+                },
+            )
+            .await
+        }
+        codex::ClientRequest::ModelList { request_id, .. } => {
+            send_typed_json_result::<codex::ModelListResponse>(
+                outbound,
+                request_id,
+                json!({
+                    "data": [model_summary(state)],
+                    "nextCursor": null
+                }),
+            )
+            .await
+        }
+        codex::ClientRequest::ModelProviderCapabilitiesRead { request_id, .. } => {
+            send_result(
+                outbound,
+                request_id,
+                codex::ModelProviderCapabilitiesReadResponse {
+                    namespace_tools: false,
+                    image_generation: false,
+                    web_search: false,
+                },
+            )
+            .await
+        }
+        codex::ClientRequest::ConfigRead { request_id, .. } => {
+            send_typed_json_result::<codex::ConfigReadResponse>(
+                outbound,
+                request_id,
+                json!({
+                    "config": {
+                        "model": state.model.as_ref(),
+                        "model_provider": "defra",
+                        "approval_policy": "never",
+                        "sandbox_mode": "danger-full-access"
+                    },
+                    "origins": {}
+                }),
+            )
+            .await
+        }
+        codex::ClientRequest::ConfigBatchWrite { request_id, .. }
+        | codex::ClientRequest::ConfigValueWrite { request_id, .. } => {
+            send_typed_json_result::<codex::ConfigWriteResponse>(
+                outbound,
+                request_id,
+                json!({
+                    "status": "ok",
+                    "version": "defra-shim",
+                    "filePath": absolute_path(&state.codex_home.join("config.toml")),
+                    "overriddenMetadata": null
+                }),
+            )
+            .await
+        }
+        codex::ClientRequest::ConfigRequirementsRead { request_id, .. } => {
+            send_result(
+                outbound,
+                request_id,
+                codex::ConfigRequirementsReadResponse { requirements: None },
+            )
+            .await
+        }
+        codex::ClientRequest::ExternalAgentConfigDetect { request_id, .. } => {
+            send_result(
+                outbound,
+                request_id,
+                codex::ExternalAgentConfigDetectResponse { items: Vec::new() },
+            )
+            .await
+        }
+        codex::ClientRequest::ExternalAgentConfigImport { request_id, .. } => {
+            send_result(
+                outbound,
+                request_id,
+                codex::ExternalAgentConfigImportResponse {},
+            )
+            .await
+        }
+        codex::ClientRequest::ExperimentalFeatureList { request_id, .. } => {
+            send_result(
+                outbound,
+                request_id,
+                codex::ExperimentalFeatureListResponse {
+                    data: Vec::new(),
+                    next_cursor: None,
+                },
+            )
+            .await
+        }
+        codex::ClientRequest::PermissionProfileList { request_id, .. } => {
+            send_result(
+                outbound,
+                request_id,
+                codex::PermissionProfileListResponse {
+                    data: Vec::new(),
+                    next_cursor: None,
+                },
+            )
+            .await
+        }
+        codex::ClientRequest::CollaborationModeList { request_id, .. } => {
+            send_result(
+                outbound,
+                request_id,
+                codex::CollaborationModeListResponse { data: Vec::new() },
+            )
+            .await
+        }
+        codex::ClientRequest::SkillsList { request_id, .. } => {
+            send_result(
+                outbound,
+                request_id,
+                codex::SkillsListResponse { data: Vec::new() },
+            )
+            .await
+        }
+        codex::ClientRequest::HooksList { request_id, .. } => {
+            send_result(
+                outbound,
+                request_id,
+                codex::HooksListResponse { data: Vec::new() },
+            )
+            .await
+        }
+        codex::ClientRequest::PluginList { request_id, .. } => {
+            send_result(
+                outbound,
+                request_id,
+                codex::PluginListResponse {
+                    marketplaces: Vec::new(),
+                    marketplace_load_errors: Vec::new(),
+                    featured_plugin_ids: Vec::new(),
+                },
+            )
+            .await
+        }
+        codex::ClientRequest::McpServerStatusList { request_id, .. } => {
+            send_result(
+                outbound,
+                request_id,
+                codex::ListMcpServerStatusResponse {
+                    data: Vec::new(),
+                    next_cursor: None,
+                },
+            )
+            .await
+        }
+        other => unreachable!(
+            "non-basic Codex request routed to basic handler: {}",
+            other.method()
+        ),
+    }
+}
